% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util_BroadBand2mun.R
\name{Util_BroadBand2mun}
\alias{Util_BroadBand2mun}
\title{Add ISTAT municipality codes to the Ultra-broadband activation DB
downloaded with \code{\link{Get_BroadBand}}}
\usage{
Util_BroadBand2mun(
  data,
  input_School2mun = NULL,
  input_Registry = NULL,
  input_AdmUnNames = NULL,
  verbose = FALSE,
  autoAbort = FALSE
)
}
\description{
Add ISTAT municipality codes to the Ultra-broadband activation DB
downloaded with \code{\link{Get_BroadBand}}
}
\keyword{internal}
